changed underlying default type to uint16_t

+++ src/Utils/UnderlyingIndexType.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
+#include <cstdint>
 namespace jGraph::internals
 {
-using underlyingGraphIndex_t = unsigned;
+using underlyingGraphIndex_t = uint16_t;
 } // namespace jGraph::internals
